+++ packages/spot-consumer/src/consumers/index.ts
@@ -1,2 +1,3 @@
 export * from './SubaccountConsumer'
+export * from './SpotMarketChronosConsumer'
 export * from './SpotMarketConsumer'