+++ packages/spot-consumer/src/index.ts
@@ -1,3 +1,4 @@
 export * from './consumers'
 export * from './transformers'
+export * from './composers'
 export * from './types'