+++ packages/spot-consumer/src/types/index.ts
@@ -4,6 +4,7 @@ import {
 } from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb'
 
 export * from './spot-market'
+export * from './spot-market-chronos'
 export * from './subaccount'
 
 export { GrpcPriceLevel, GrpcCosmosCoin }