+++ packages/exchange-consumer/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.30](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exchange-consumer@0.0.29...@injectivelabs/exchange-consumer@0.0.30) (2021-04-16)
+
+### Features
+
+* added new composer to spot-consumer ([f13e4ef](https://github.com/InjectiveLabs/injective-ts/commit/f13e4efcabde66ada37549531220692769ac0e45))
+
 ## [0.0.29](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exchange-consumer@0.0.28...@injectivelabs/exchange-consumer@0.0.29) (2021-04-10)
 
 ### Features