+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.10](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.9...@injectivelabs/spot-consumer@0.0.10) (2021-04-16)
+
+### Features
+
+* added new composer to spot-consumer ([f13e4ef](https://github.com/InjectiveLabs/injective-ts/commit/f13e4efcabde66ada37549531220692769ac0e45))
+
 ## [0.0.9](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.8...@injectivelabs/spot-consumer@0.0.9) (2021-04-14)
 
 **Note:** Version bump only for package @injectivelabs/spot-consumer